App Server thread mutations
